Refactor to use a common component
